SQL services - interfaces

+++ persistence/src/main/java/cz/muni/fi/pa036/dao/RestaurantPersistenceDao.java
@@ -27,9 +27,8 @@ public interface RestaurantPersistenceDao {
 
     
     Restaurant findByName(String name);
-
     
-    Restaurant findById(Long id);
+    Restaurant findById(int id);
     
     double getRating(String restaurantId);
